canCalculateMarkup passes for piano

+++ src/main/java/behaviours/ISell.java
@@ -2,5 +2,5 @@ package behaviours;
 
 public interface ISell {
 
-    int calculateMarkup();
+    double calculateMarkup();
 }
